requirements.txt
streamlit==1.22.0
crewai==0.28.8
crewai_tools==0.1.6
langchain_community==0.0.29
duckduckgo-search==6.2.11
PyPDF2==3.0.1
openai==1.42.0